Implement state machine and maintenance mode
Implement state machine to determine which inputs and outputs are
relevant and which actions are possible at each moment, and implement
maintenance mode which allows plunger to be moved in and out for
cleaning. Uses a simple error handler to resolve task errors and axis
faults. Verified to compile and run on Soloist CP.