Add sinusoidal motion to experiment mode
Replace constant velocity motion with sinusoidal motion in experiment
mode. Sinusoidal motion is computed using a cubic-spline interpolation
(PVT) on sampled points from a model of the motion. Note that some
debugging code was left commented in the code for future use. Verified
to run and compile on Soloist CP.