Add functions to rotate and reflect integers
Add helper functions to rotate left, rotate right, and reflect integers.
Functions will be used when manipulating data shifted into controller
from parameter read circuitry or shifted out of controller to parameter
display circuitry. Note that code was originally developed for use on
unsigned integers in C but should be compatible with signed integers in
Aerobasic. Not compiled or tested in any way on Soloist CP.