Add functions to format and print to SSD
Add functions to format numbers to be printed on a seven-segment display
and to print numbers to a pair of seven-segment displays. The pair of
seven-segment displays uses a simple interface circuit consisting of an
8-bit shift register, an 8-bit register, and two BCD-to-SSD drivers with
independent clocks for the shift register and the register. Not compiled
or tested in any way on Soloist CP.